% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_api_stats}
\alias{get_api_stats}
\title{Function to get account statistics from Gender API v2}
\usage{
get_api_stats(api_token)
}
\arguments{
\item{api_token}{your API token as a string (Bearer token)}
}
\value{
a data frame with account statistics
}
\description{
This function retrieves account usage statistics
}
