% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_country_codes}
\alias{get_country_codes}
\title{Helper function to create country code reference}
\usage{
get_country_codes()
}
\value{
a data frame with country codes, names, and regions
}
\description{
This function returns a data frame with all supported country codes
}
