% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_gender}
\alias{get_gender}
\title{Function to query Gender API}
\usage{
get_gender(name, api_key, country_code = NA)
}
\arguments{
\item{name}{desired given name as  a string}

\item{api_key}{your pai key as a string}

\item{country_code}{optional country code (e.g. "US")}
}
\value{
a data table with prediction of gender
}
\description{
This function uses Gender API to predict gender based on first name and country if desired
}
