% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_gender_batch}
\alias{get_gender_batch}
\title{Function to process multiple names in batch}
\usage{
get_gender_batch(names_list, api_token, request_type = "first_name")
}
\arguments{
\item{names_list}{a list of lists, each containing name information (first_name, country, etc.)}

\item{api_token}{your API token as a string (Bearer token)}

\item{request_type}{type of request: "first_name", "full_name", or "email"}
}
\value{
a data frame with results for all names
}
\description{
This function allows batch processing of multiple names
}
