% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{handle_api_errors}
\alias{handle_api_errors}
\title{Helper function for error handling}
\usage{
handle_api_errors(response)
}
\arguments{
\item{response}{httr response object}
}
\value{
processed error message or NULL if no error
}
\description{
This function processes API error responses
}
